+++ Code/Controllers/ATLConversationListViewController.h
@@ -260,10 +260,7 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  @abstract The controller used to display search results.
  */
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
-@property (nonatomic, readonly) UISearchDisplayController *searchController;
-#pragma GCC diagnostic pop
+@property (nonatomic, readonly) UISearchController *searchController;
 
 /**
  @abstract A boolean value that determines if the controller should show a search bar and search display controller.